backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-options --output-file=requirements.txt requirements.in
#
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
bcrypt==3.2.2
    # via -r requirements.in
cffi==1.15.0
    # via bcrypt
click==8.0.4
    # via uvicorn
fastapi==0.78.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
passlib==1.7.4
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via fastapi
pyjwt==2.4.0
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
six==1.16.0
    # via python-multipart
sniffio==1.2.0
    # via anyio
starlette==0.19.1
    # via fastapi
typing-extensions==4.1.1
    # via pydantic
uvicorn[standard]==0.17.6
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
